# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Aiexec
  version: 1.6.4
paths:
...
  /api/v1/knowledge_bases/:
    get:
      tags:
      - Knowledge Bases
      summary: List Knowledge Bases
      description: List all available knowledge bases.
      operationId: list_knowledge_bases_api_v1_knowledge_bases__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/KnowledgeBaseInfo'
                type: array
                title: Response List Knowledge Bases Api V1 Knowledge Bases  Get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    delete:
      tags:
      - Knowledge Bases
      summary: Delete Knowledge Bases Bulk
      description: Delete multiple knowledge bases.
      operationId: delete_knowledge_bases_bulk_api_v1_knowledge_bases__delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Delete Knowledge Bases Bulk Api V1 Knowledge Bases  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v1/knowledge_bases:
    get:
      tags:
      - Knowledge Bases
      summary: List Knowledge Bases
      description: List all available knowledge bases.
      operationId: list_knowledge_bases_api_v1_knowledge_bases_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/KnowledgeBaseInfo'
                type: array
                title: Response List Knowledge Bases Api V1 Knowledge Bases Get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    delete:
      tags:
      - Knowledge Bases
      summary: Delete Knowledge Bases Bulk
      description: Delete multiple knowledge bases.
      operationId: delete_knowledge_bases_bulk_api_v1_knowledge_bases_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Delete Knowledge Bases Bulk Api V1 Knowledge Bases
                  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v1/knowledge_bases/{kb_name}:
    get:
      tags:
      - Knowledge Bases
      summary: Get Knowledge Base
      description: Get detailed information about a specific knowledge base.
      operationId: get_knowledge_base_api_v1_knowledge_bases__kb_name__get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: kb_name
        in: path
        required: true
        schema:
          type: string
          title: Kb Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeBaseInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Knowledge Bases
      summary: Delete Knowledge Base
      description: Delete a specific knowledge base.
      operationId: delete_knowledge_base_api_v1_knowledge_bases__kb_name__delete
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: kb_name
        in: path
        required: true
        schema:
          type: string
          title: Kb Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Delete Knowledge Base Api V1 Knowledge Bases  Kb Name  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BulkDeleteRequest:
      properties:
        kb_names:
          items:
            type: string
          type: array
          title: Kb Names
      type: object
      required:
      - kb_names
      title: BulkDeleteRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KnowledgeBaseInfo:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        embedding_provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Embedding Provider
          default: Unknown
        embedding_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Embedding Model
          default: Unknown
        size:
          type: integer
          title: Size
          default: 0
        words:
          type: integer
          title: Words
          default: 0
        characters:
          type: integer
          title: Characters
          default: 0
        chunks:
          type: integer
          title: Chunks
          default: 0
        avg_chunk_size:
          type: number
          title: Avg Chunk Size
          default: 0.0
      type: object
      required:
      - id
      - name
      title: KnowledgeBaseInfo
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/login
    API key query:
      type: apiKey
      in: query
      name: x-api-key
    API key header:
      type: apiKey
      in: header
      name: x-api-key